salesforce_streamlit_app/requirements.txt
streamlit>=1.32.0
requests>=2.31.0
pandas>=2.2.0
altair>=5.2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
simple-salesforce>=1.12.5